requirements.txt
telethon
python-telegram-bot
requests
beautifulsoup4
schedule
lxml
